fix: rename api to api-calls
