+++ components/forms/signup-form.tsx
@@ -17,7 +17,7 @@ import { Input } from '@/components/ui/input';
 
 import { Icons } from '@/components/icons';
 
-import { signUp } from '@/api/auth';
+import { signUp } from '@/api-calls/auth';
 
 import { useFrequentlyUsedHooks } from '@/hooks/use-frequently-used-hooks';
 import { useFocus } from '@/hooks/use-focus';
